Nope, didn't work.
Appending all those nodes to the DOM takes several minutes even if the table is `display: none`.
When the table is made visible, it hangs the page for at least several seconds.

On a ~2GB, ~1M-entry log file, making the table visible crashes the tab with an OOM error, even though none of that memory usage is on the JS heap.
It seems like the tab takes >2GB of memory even though the original file is only 2GB big.

This reverts commits 621c6502029330397de1f0549eb1e350c95ae379 and 986a09ac4cd8af86944ef2a3b60e5be6dd78a1af.